Overhaul to add react-markdown.
All pages are now defined as a component and markdown. Content is provided in the markdown file.
