+++ src/pages/index/index.md
@@ -1,0 +1,16 @@
+## <span class=yellow>Homepage</span>
+
+### <span class="nav_link yellow">[View the #seasonATE page](/seasonate)</span>
+
+### <span class="nav_link yellow">[View the #SNACKRIFICE page](/snackrifice)</span>
+
+### <span class="nav_link yellow">[View team resources](/resources)</span>
+
+<div class="img-box">
+
+![This taco is under construction, check back soon.](https://media.discordapp.net/attachments/755560065645281360/755577280989888562/tacoconstruction.png)
+
+</div>
+
+This site is in progress. Stay tuned!
+
