+++ src/pages/resources/resources.md
@@ -1,0 +1,6 @@
+## Team Resources
+
+- <span class="nav_link yellow">[The Unlimited Taco Stand (Discord)](https://discord.gg/fm6mRwy)</span>
+- <span class="nav_link yellow">[The Menu](https://docs.google.com/spreadsheets/d/1BYOS1OvOxx3_9fg5t1UBHRXFastTxJmnP7xWpHr3YNU/edit#gid=935168686)</span>
+- <span class="nav_link yellow">[Welcome Presentation](https://drive.google.com/file/d/1zClARg9yLVn6IWD4Gb7a_0HgiIvE1jEh/view)</span>
+- <span class="nav_link yellow">[Strategy and coordination](https://docs.google.com/spreadsheets/d/1uC3fFzRmxWPQt8maFVWsDHo1kACiaM09R9CN9YUl5YU/edit#gid=1688490183)</span>
