+++ src/pages/seasonate/seasonate.md
@@ -1,0 +1,21 @@
+## <span class=yellow>#seasonATE</span>
+
+<div class="img-box">
+
+![seasonATE announcement, contents in plain text below](https://cdn.discordapp.com/attachments/734574989457621034/757708864111247461/image0.png)
+
+</div>
+
+Hi there, Tacos here!
+
+First off, thank you EVERYONE for making this goal of ours last season come true. It started as a way to demonstrate to The Peanut that they aren't our boss and that still remains true. Also, to quote one of the keepers, a commitment to beta-testing
+
+We would like to just put forward a few things which we have established about our brand new (slightly used) pitching machine and how it relates to last season's <span class="yellow">#SNACKRIFICE</span>
+
+1) The Tacos aren't striking against the ILB nor the commissioner. Pitching Machine is a wonderful compromise between the Tacos and the league to continue to play ball while respecting the <span class="nav_link yellow">#SNACKRIFICE</span>
+
+2) We LOVE the pitching machine. Defiance to the Gods is not the same as defiance to the ILB. We would appreciate it if our lovely pitching machine (with its "This machine kills capitalism" sticker) is not referred to as a scab. It didn't ask for this and, well until further notice, it doesn't really ask. We are unsure of its sentience
+
+3) The commissioner is doing a great job
+
+From your Los Angeli Unlimited Tacos
