+++ src/pages/snackrifice/snackrifice.md
@@ -1,0 +1,117 @@
+## <span class="yellow">#SNACKRIFICE</span>
+          
+### Thank you for your support!
+
+<div class="img-box">
+
+![Thank you note, contents below](https://cdn.discordapp.com/attachments/734574989457621034/757365017787236382/og9NipC.png)
+
+</div>
+
+Hello everyone,
+
+This is a message from the Tacos to you, yes you, the person reading this.
+
+First, thank you. The Community has proved with this success that we are capable of coming together and doing great things in the name of Blaseball, and we hope that willingness to band together continues. Splortsmanship is an important trait for the Tacos, and in this event we have ALL shown a willingness to come together under one goal, to unite, to rise up as one powerful voice and accomplish a monumental, nearly impossible task.
+
+The Tacos aren’t sure what’s going to happen tomorrow. Nothing has been revealed to us in any way, we can only wonder the same as everyone else. Regardless of the outcome, we are grateful to have completed this alongside you. This success is your success. 
+
+Today we have made Blaseball history, not alone, but together, as a diverse and dedicated group of splorts fans. 
+
+<span class="yellow">WE</span> did it. 
+
+<span class="yellow">WE</span> won. 
+
+All of us, together, made the Snackrifice Complete. 
+
+Despite all else that has happened today, be proud. 
+
+We are all love Blaseball.
+
+Love,
+
+<span class="yellow">The Unlimited Tacos.</span>
+
+### <span class="nav_link yellow">[See #seasonATE](/seasonate)</span>
+
+<div class="img-box">
+
+![Leaderboard after #SNACKRIFICE showing all pitchers shelled.](https://cdn.discordapp.com/attachments/755560065645281360/756633189958549544/unknown.png)
+
+Thanks to your efforts, the Tacos were able to sacrifice our pitchers and save your players. We were also able to bring our challenge directly to the blaseball gods.
+
+### <span class="yellow">Stage 2:</span> Drop the Tacos
+
+It's been a wild ride, but we still need your help.
+
+Now that our pitchers are shelled, it is important to preserve them for season 8. If you are idolizing them, stop now. We need them out of positions 14 and 21. We cannot risk blessing messing this up.
+
+
+### The Legume Descends
+
+<div class="vid-box">
+<iframe width="100%" height="100%" src="https://www.youtube.com/embed/KhFnT42anvU" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen title="The Peanut Arrives"></iframe>
+</div>
+
+### <span class="yellow">Stage 1:</span> Save your players, Shell our pitchers.
+
+Listen, we know you love getting filthy rich off of Axel Trololol or PolkaDot Patterson, but do you want them to be trapped in a peanut shell for the next season?
+
+#### Here's the Plan
+Get every single Tacos pitcher up to the top 10, above the red line. This will shell all of our pitchers (A natural evolution for a Taco) and save the other players.
+</div>
+
+<div class="img-box">
+
+![Tacos belong in shells, your players don't. Idolize one of our pitchers before the season ends.](https://media.discordapp.net/attachments/734574989457621034/755809539974103050/Tacos-belong-in-shells-PixTeller.png)
+
+</div>
+
+<Leaderboard />
+
+If you're not sure who to support, we have an easy guide that will match you with one of our lovely pitchers so you can give them the love and support they need.
+
+<span>
+
+<a id="pitcher_matcher"
+href="https://media.discordapp.net/attachments/755560065645281360/755568581374378026/tacolist2.png" target="new">
+&nbsp;🌮&nbsp;Get&nbsp;Matched&nbsp;Here&nbsp;</a>
+</span>
+
+So what are you waiting for? Defy the gods!
+
+<div class="img-box">
+
+![Going on siesta? Have a taco! Idolize a tacos pitcher before the season ends.](https://media.discordapp.net/attachments/734574989457621034/755803529251258390/Going-on-siesta-PixTeller_2.png)
+
+</div>
+
+<details>
+
+<summary>Click here to view some infographics.</summary>
+
+<div class="img-box">
+
+![Infographic explaining the snackrifice.](https://media.discordapp.net/attachments/755560065645281360/755568614681083994/Our_Snackrifice.png)
+
+</div>
+
+<div class="img-box">
+
+![Idol a taco, Support the snackrifice.](https://media.discordapp.net/attachments/755560065645281360/755568625804378312/Idol-A-Taco.png)
+
+</div>
+
+<div class="img-box">
+
+![It's a beautiful dat for a #snackrifice. 72 degrees and infinites.](https://media.discordapp.net/attachments/755560065645281360/755568676630954044/snackrificesunrise.png)
+
+</div>
+
+<div class="img-box">
+
+![It's day 80, have you shelled your pitcher? Save your star, Shell a taco today!](https://media.discordapp.net/attachments/734574989457621034/755801659963080865/day_80.png)
+
+</div>
+
+</details>
