+++ src/pages/template/template.md
@@ -1,0 +1,7 @@
+## This is a template
+
+Put things here.
+
+Markdown format.
+
+We allow HTML, so feel free to <span class="yellow">add things if you need to.</span>
